dash-tui: split out the tui code
